dev environment: Use development source when webui is present
When installing FreeIPA with support for the modern webui (e.g. Fedora
43, NVR 4.13.2-2) from the distribution packages, the webui entry for
the package takes precedence over the development configuration, causing
the pages to seem to not update when using a browser.

By either modifying the modern webui path, if it is intalled, or by
adding it otherwise, we ensure that the development version is always
the one used during development.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>